requirements.txt
openai
mistralai
nltk
textblob
transformers
sentence-transformers
torch
python-dotenv
requests
